separate hardhat out from defender action code
