+++ contracts/scripts/defender-actions/rollup.config.cjs
@@ -28,21 +28,20 @@ const commonConfig = {
     "ethers",
     "web3",
     "mocha",
-    "hardhat",
   ],
 };
 
 module.exports = [
-  {
+   {
     ...commonConfig,
-    input: "registerValidators.js",
+    input: "doAccounting.js",
     output: {
-      file: "dist/registerValidators/index.js",
+      file: "dist/doAccounting/index.js",
       inlineDynamicImports: true,
       format: "cjs",
     },
   },
-  {
+  { 
     ...commonConfig,
     input: "stakeValidators.js",
     output: {
@@ -53,9 +52,9 @@ module.exports = [
   },
   {
     ...commonConfig,
-    input: "doAccounting.js",
+    input: "registerValidators.js",
     output: {
-      file: "dist/doAccounting/index.js",
+      file: "dist/registerValidators/index.js",
       inlineDynamicImports: true,
       format: "cjs",
     },
