+++ contracts/utils/signers.js
@@ -1,11 +1,7 @@
 const { Wallet } = require("ethers");
 const { parseEther } = require("ethers/lib/utils");
-const {
-  DefenderRelaySigner,
-  DefenderRelayProvider,
-} = require("@openzeppelin/defender-relay-client/lib/ethers");
 const hhHelpers = require("@nomicfoundation/hardhat-network-helpers");
-
+const { getDefenderSigner } = require("./signersNoHardhat");
 const { ethereumAddress, privateKey } = require("./regex");
 
 const log = require("./logger")("utils:signers");
@@ -60,39 +56,6 @@ async function getSigner(address = undefined) {
 
   return signer;
 }
-
-const getDefenderSigner = async () => {
-  const speed = process.env.SPEED || "fastest";
-  if (!["safeLow", "average", "fast", "fastest"].includes(speed)) {
-    console.error(
-      `Defender Relay Speed param must be either 'safeLow', 'average', 'fast' or 'fastest'. Not "${speed}"`
-    );
-    process.exit(2);
-  }
-
-  const { chainId } = await ethers.provider.getNetwork();
-  const isMainnet = chainId === 1;
-
-  const apiKey = isMainnet
-    ? process.env.DEFENDER_API_KEY
-    : process.env.HOLESKY_DEFENDER_API_KEY || process.env.DEFENDER_API_KEY;
-  const apiSecret = isMainnet
-    ? process.env.DEFENDER_API_SECRET
-    : process.env.HOLESKY_DEFENDER_API_SECRET ||
-      process.env.DEFENDER_API_SECRET;
-
-  const credentials = { apiKey, apiSecret };
-
-  const provider = new DefenderRelayProvider(credentials);
-  const signer = new DefenderRelaySigner(credentials, provider, {
-    speed,
-  });
-
-  log(
-    `Using Defender Relayer account ${await signer.getAddress()} with key ${apiKey} and speed ${speed}`
-  );
-  return signer;
-};
 
 /**
  * Impersonate an account when connecting to a forked node.
